LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-mouse-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<p>
Please run this with DRT or WTR.
</p>
Test following mouse actions:
<ul>
    <li>Mouse click to focus each of sub-fields</li>
    <li>Mouse click on the spin button to update each of sub-fields</li>
</ul>
<input id="input" type="date" style="font-family:ahem; font-size:16px;">
<div id="console"></div>
<script>
description('Multiple fields UI of date input type with mouse events');
var input = document.getElementById('input');

function keyDown(key, modifiers)
{
    if (!window.eventSender)
        return;
    eventSender.keyDown(key, modifiers);
}

function mouseClickOn(x, y)
{
    if (!window.eventSender)
        return;
    eventSender.mouseMoveTo(x + input.offsetLeft, y + input.offsetTop);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

input.value = '2345-07-19';
var center = input.offsetHeight / 2;
var spinButtonOffset = 26;

debug('==> Focus on the month field.');
mouseClickOn(12, center);
keyDown('upArrow');
shouldBeEqualToString('input.value', '2345-08-19');
mouseClickOn(input.offsetWidth - spinButtonOffset, center - 1);
shouldBeEqualToString('input.value', '2345-09-19');
mouseClickOn(input.offsetWidth - spinButtonOffset, center + 1);
shouldBeEqualToString('input.value', '2345-08-19');
shouldBeZero('window.getSelection().rangeCount'); // No text selection.

debug('');
debug('==> Focus on the day field.');
mouseClickOn(60, center);
keyDown('upArrow');
shouldBeEqualToString('input.value', '2345-08-20');
mouseClickOn(input.offsetWidth - spinButtonOffset, center - 1);
shouldBeEqualToString('input.value', '2345-08-21');
mouseClickOn(input.offsetWidth - spinButtonOffset, center + 1);
shouldBeEqualToString('input.value', '2345-08-20');
shouldBeZero('window.getSelection().rangeCount'); // No text selection.

debug('');
debug('==> Focus on the year field.');
mouseClickOn(input.offsetWidth - 100, center);
keyDown('upArrow');
shouldBeEqualToString('input.value', '2346-08-20');
mouseClickOn(input.offsetWidth - spinButtonOffset, center - 1);
shouldBeEqualToString('input.value', '2347-08-20');
mouseClickOn(input.offsetWidth - spinButtonOffset, center + 1);
shouldBeEqualToString('input.value', '2346-08-20');
shouldBeZero('window.getSelection().rangeCount'); // No text selection.

debug('');
debug('==> Click on a disabled field.');
input.disabled = true;
mouseClickOn(12, center);
keyDown('upArrow');
shouldBeEqualToString('input.value', '2346-08-20');
input.disabled = false;

debug('');
debug('==> Click on a read-only field.');
input.readOnly = true;
mouseClickOn(12, center);
keyDown('upArrow');
shouldBeEqualToString('input.value', '2346-08-20');
input.readOnly = false;

debug('');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
